Refactor product-related classes to use Integer for quantity and update purchase endpoint

+++ services/product/src/main/java/com/jn/ecommerce/product/ProductMapper.java
@@ -32,7 +32,7 @@ public class ProductMapper {
                 .build();
     }
 
-    public ProductPurchaseResponse toProductPurchaseResponse(Product product,  double quantity) {
+    public ProductPurchaseResponse toProductPurchaseResponse(Product product,  Integer quantity) {
         return new ProductPurchaseResponse(
                 product.getId(),
                 product.getName(),
