+++ services/product/src/main/java/com/jn/ecommerce/product/ProductResponse.java
@@ -10,7 +10,7 @@ public record ProductResponse(
         String name,
         String description,
         BigDecimal price,
-        double availableQuantity,
+        Integer availableQuantity,
         Integer categoryId,
         String categoryName
 ) {
